Refactor plugin system for performance
