--- src/extend/PluginContext.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const PluginContext = React.createContext();
-
-export default PluginContext;
