+++ src/lib/MiradorViewer.js
@@ -2,12 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import deepmerge from 'deepmerge';
-import PluginProvider from '../extend/PluginProvider';
-import App from '../containers/App';
+import { storePlugins } from '../extend/pluginStore';
 import createStore from '../state/createStore';
-import createRootReducer from '../state/reducers/rootReducer';
 import * as actions from '../state/actions';
 import settings from '../config/settings';
+import App from '../containers/App';
 
 /**
  * Default Mirador instantiation
@@ -16,7 +15,9 @@ class MiradorViewer {
   /**
    */
   constructor(config, plugins) {
-    this.store = createStore();
+    storePlugins(plugins);
+    const pluginReducers = getReducersFromPlugins(plugins);
+    this.store = createStore(pluginReducers);
     this.config = config;
     this.processConfig();
     const viewer = {
@@ -26,9 +27,7 @@ class MiradorViewer {
 
     ReactDOM.render(
       <Provider store={this.store}>
-        <PluginProvider plugins={plugins} createRootReducer={createRootReducer}>
-          <App />
-        </PluginProvider>
+        <App />
       </Provider>,
       document.getElementById(config.id),
     );
@@ -72,6 +71,11 @@ class MiradorViewer {
       );
     });
   }
+}
+
+/** Return reducers from plugins */
+function getReducersFromPlugins(plugins) {
+  return plugins && plugins.reduce((acc, plugin) => ({ ...acc, ...plugin.reducers }), {});
 }
 
 export default MiradorViewer;
